requirements.txt
Django==2.2.4
django-rest-auth==0.9.5
djangorestframework==3.10.2
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
django-allauth==0.40.0